reorganizing files to be published/replaced
