+++ packages/oc-template-react-compiler/lib/compile.js
@@ -3,7 +3,7 @@
 const _ = require("lodash");
 const async = require("async");
 const compileServer = require("./compileServer");
-const compileStatics = require("./compileStatics");
+const compileStatics = require("./to-be-published/compileStatics");
 const compileView = require("./compileView");
 const fs = require("fs-extra");
 const getInfo = require("oc-template-react").getInfo;
