+++ packages/oc-template-react-compiler/lib/compileServer.js
@@ -1,7 +1,6 @@
 "use strict";
 
 const async = require("async");
-const compiler = require("./compiler");
 const fs = require("fs-extra");
 const hashBuilder = require("oc-hash-builder");
 const path = require("path");
@@ -9,7 +8,9 @@ const path = require("path");
 // TODO:
 // Abstract oc-webpack.configurator to support dev/publish scenario
 // It should enable what we are doing here (configurator and compiler)
-const { webpackConfigurator } = require("./to-be-published");
+const compiler = require("./to-abstract-base-template-utils/compiler");
+const webpackConfigurator = require("./to-abstract-base-template-utils/webpackConfigurator");
+const reactComponentWrapper = require("./to-be-published/oc-react-component-wrapper");
 
 module.exports = ({ options, compiledInfo }, callback) => {
   const serverFileName = options.componentPackage.oc.files.data;
