+++ packages/oc-template-react-compiler/lib/to-abstract-base-template-utils/webpackConfigurator.js
