+++ packages/oc-template-react-compiler/lib/to-be-published/compileStatics.js
