--- packages/oc-template-react-compiler/lib/to-be-published/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  webpackConfigurator: require("./webpackConfigurator"),
-  reactComponentWrapper: require("./oc-react-component-wrapper")
-};
